task#10477 Worked more on this in the context of the videolearn project.
Worked on the FieldType and events action related to them.
Added a static level variable to the FieldType so the mydb_events level and incrementally generated to avoid any conflict.

Tried to push it to reference individual fields on the eventaction with the associated registry information.
Got into issue for referencing / naming the event and retrieving the registry values.
Think of using the dataobject persistance once I've found how to pass the session object name to the FieldType.

As of now most serverside validation is broken, will fix it soon.